Developer Environment -- node:alpine dockerfile
The purpose of this repo is to provide tools needed for a standard development environment.

A standard environment will make setup instructions and troubleshooting easier across
disparate platforms.

This Dockerfile provides node and npm with alpine linux. No project dependencies are included
to keep it lightweight. It's also good boilerplate for project specific Dockerfiles